# perf: improve getHotelsByDestinationId api average latency