Load testing is run using k6.

**Before Commmit**

Config: 30VUS, Duration: 10s
<img width="811" alt="30vus" src="https://user-images.githubusercontent.com/13248466/222538842-ce1e9c60-a229-4a03-84ad-6fa36c1028cf.png">

Config: 100VUs, Duration: 10s
<img width="795" alt="100vus" src="https://user-images.githubusercontent.com/13248466/222538859-4d0b5073-ccf8-4bb3-b1df-b05f8bcaf833.png">

**After Commit:**

Config: 30VUS, Duration: 10s
<img width="827" alt="new30vus" src="https://user-images.githubusercontent.com/13248466/222538975-e3d86ec0-dc76-45de-8b25-fb6439f1cba5.png">

Config: 100VUS, Duration: 10s
<img width="825" alt="new100vus" src="https://user-images.githubusercontent.com/13248466/222539035-0714022a-b07d-40b9-ab44-4837d3d7c329.png">